backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.54.0
python-dotenv==1.0.0
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
dnspython==2.4.2
